refactor: move ExponentialBackoff to internal
