+++ clients/src/main/java/org/apache/kafka/clients/ClusterConnectionStates.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.clients;
 
 import org.apache.kafka.common.errors.AuthenticationException;
-import org.apache.kafka.common.utils.ExponentialBackoff;
+import org.apache.kafka.common.utils.internals.ExponentialBackoff;
 import org.apache.kafka.common.utils.LogContext;
 
 import org.slf4j.Logger;
