+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/RequestState.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.clients.consumer.internals;
 
-import org.apache.kafka.common.utils.ExponentialBackoff;
+import org.apache.kafka.common.utils.internals.ExponentialBackoff;
 import org.apache.kafka.common.utils.LogContext;
 
 import org.slf4j.Logger;
