+++ clients/src/main/java/org/apache/kafka/common/utils/internals/ExponentialBackoff.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.common.utils;
+package org.apache.kafka.common.utils.internals;
 
 import java.util.concurrent.ThreadLocalRandom;
